Configure nvr and ranger support
nvr can be used to open files, run vim commands, and more from multiple
terminals, all directed at a single remote instance of neovim.

One example of this functionality is the ability to configure ranger
shortcuts that run nvr to open files in different splits or tabs.
.selected_editor determines the command to run when simply hitting enter
within ranger (open with editor) in the few tests I've run. Likely, it
will not work for all filetypes. The shortcuts for splits and tabs are
in rc.conf.